(#59) (#83) (#89) 운동기록캘린더 개발, 소셜회원 통합에따른 재로그인 안내로직 개발, 각종 약관 및 구글폼 연결 수정내용: weeks생성부분 exerciseCalendar에서 monthView로 이동
