+++ src/app/exercise/_components/monthView.tsx
@@ -1,8 +1,18 @@
 'use client'
 
-import { memo, useState, useEffect } from 'react'
+import { memo, useState, useEffect, useMemo } from 'react'
 import { useRouter } from 'next/navigation'
-import { format, isSameMonth, isSameDay, isAfter } from 'date-fns'
+import {
+  format,
+  isSameMonth,
+  isSameDay,
+  isAfter,
+  startOfWeek,
+  endOfWeek,
+  startOfMonth,
+  endOfMonth,
+  addDays,
+} from 'date-fns'
 import clsx from 'clsx'
 
 import StampIcon from '@/assets/stamp.svg'
@@ -11,20 +21,31 @@ import Typography from '@/components/ui/typography'
 import styles from './monthView.module.css'
 
 type Props = {
-  weeks: Date[][]
   month: Date
   today: Date
   counts: Record<string, number>
 }
 
-const MonthView = memo(function MonthView({
-  weeks,
-  month,
-  today,
-  counts,
-}: Props) {
+const MonthView = memo(function MonthView({ month, today, counts }: Props) {
   const router = useRouter()
   const [notCacheData, setNotCacheData] = useState(false)
+
+  const weeks = useMemo(() => {
+    const gridStart = startOfWeek(startOfMonth(month), { weekStartsOn: 0 })
+    const gridEnd = endOfWeek(endOfMonth(month), { weekStartsOn: 0 })
+
+    const newWeeks: Date[][] = []
+    let cursor = gridStart
+    while (cursor <= gridEnd) {
+      const days: Date[] = []
+      for (let i = 0; i < 7; i++) {
+        days.push(cursor)
+        cursor = addDays(cursor, 1)
+      }
+      newWeeks.push(days)
+    }
+    return newWeeks
+  }, [month])
 
   // counts값이 변경될때마다 캐시된 데이터인지 여부를 판단하여 스탬프 아이콘에 fadeIn 효과를 줄지 즉시 표시해줄지 판단합니다.
   useEffect(() => {
